+++ tests/beaker_tests/cisco_hsrp/test_interface_hsrp.rb
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2015 Cisco and/or its affiliates.
+# Copyright (c) 2016 Cisco and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,84 +21,92 @@
 #  - A description of the 'tests' hash and its usage
 #
 ###############################################################################
+
 require File.expand_path('../../lib/utilitylib.rb', __FILE__)
 
 # Test hash top-level keys
 tests = {
-  agent:            agent,
-  master:           master,
-  operating_system: 'nexus',
-  platform:         'n(5|6|7|9)k',
-  resource_name:    'cisco_vxlan_vtep',
+  master:        master,
+  agent:         agent,
+  intf_type:     'port-channel',
+  platform:      'n(3|7|9)k',
+  resource_name: 'cisco_interface',
 }
 
 # Skip -ALL- tests if a top-level platform/os key exludes this platform
 skip_unless_supported(tests)
+skip_nexus_image('D1', tests)
+
+# Find a usable interface for this test
+@intf = 'port-channel100'
 
 # Test hash test cases
 tests[:default] = {
-  title_pattern:  'nve1',
-  manifest_props: {
-    description:                     'default',
-    host_reachability:               'default',
-    shutdown:                        'default',
-    source_interface:                'default',
-    source_interface_hold_down_time: 'default',
+  desc:               '1.1 Default properties',
+  title_pattern:      @intf,
+  sys_def_switchport: false,
+  manifest_props:     {
+    switchport_mode:    'disabled',
+    hsrp_bfd:           'default',
+    hsrp_delay_minimum: 'default',
+    hsrp_delay_reload:  'default',
+    hsrp_mac_refresh:   'default',
+    hsrp_use_bia:       'default',
+    hsrp_version:       'default',
   },
-  resource:       {
-    'host_reachability' => 'flood',
-    'shutdown'          => 'true',
+  code:               [0, 2],
+  resource:           {
+    hsrp_bfd:           'false',
+    hsrp_delay_minimum: 0,
+    hsrp_delay_reload:  0,
+    hsrp_mac_refresh:   'false',
+    hsrp_use_bia:       'false',
+    hsrp_version:       1,
   },
 }
 
 tests[:non_default] = {
-  title_pattern:  'nve1',
+  desc:           '2.1 Non Default properties',
+  title_pattern:  @intf,
   manifest_props: {
-    description:                     'Puppet test',
-    host_reachability:               'evpn',
-    shutdown:                        'false',
-    source_interface:                'loopback55',
-    source_interface_hold_down_time: '100',
+    switchport_mode:    'disabled',
+    hsrp_bfd:           'true',
+    hsrp_delay_minimum: 100,
+    hsrp_delay_reload:  200,
+    hsrp_mac_refresh:   350,
+    hsrp_use_bia:       'use_bia_intf',
+    hsrp_version:       2,
   },
 }
 
-def unsupported_properties(*)
-  unprops = []
-  unprops << :source_interface_hold_down_time unless platform[/n(9)k/]
-  unprops
-end
-
-# Overridden to properly handle dependencies for this test file.
-def test_harness_dependencies(*)
-  return unless platform[/n(5|6)k/]
-  skip_if_nv_overlay_rejected(agent)
-
-  # Vxlan has a hard requirement to disable feature fabricpath on n5/6k
-  cmd = 'no feature-set fabricpath'
-  command_config(agent, cmd, cmd)
+def cleanup(agent)
+  cmd = 'no feature hsrp'
+  test_set(agent, cmd)
+  interface_cleanup(agent, @intf)
 end
 
 #################################################################
 # TEST CASE EXECUTION
 #################################################################
 test_name "TestCase :: #{tests[:resource_name]}" do
-  teardown do
-    resource_absent_cleanup(agent, 'cisco_vxlan_vtep')
-    vdc_limit_f3_no_intf_needed(:clear)
-  end
-  resource_absent_cleanup(agent, 'cisco_vxlan_vtep')
-  vdc_limit_f3_no_intf_needed(:set)
+  teardown { cleanup(agent) }
+  cleanup(agent)
 
-  # -----------------------------------
+  # -------------------------------------------------------------------
   logger.info("\n#{'-' * 60}\nSection 1. Default Property Testing")
-  id = :default
-  test_harness_run(tests, id)
 
+  test_harness_run(tests, :default)
+
+  id = :default
+  tests[id][:desc] = '1.2 Common Defaults (absent)'
   tests[id][:ensure] = :absent
   test_harness_run(tests, id)
 
-  # -----------------------------------
+  # -------------------------------------------------------------------
   logger.info("\n#{'-' * 60}\nSection 2. Non Default Property Testing")
+
   test_harness_run(tests, :non_default)
+  # -------------------------------------------------------------------
+  skipped_tests_summary(tests)
 end
 logger.info("TestCase :: #{tests[:resource_name]} :: End")