+++ tests/beaker_tests/cisco_overlay_global/test_overlay_global.rb
@@ -26,7 +26,7 @@ require File.expand_path('../../lib/utilitylib.rb', __FILE__)
 tests = {
   agent:         agent,
   master:        master,
-  platform:      'n(5|6|7|8|9)k',
+  platform:      'n(5|6|7|9)k',
   resource_name: 'cisco_overlay_global',
   ensurable:     false,
 }