+++ tests/beaker_tests/cisco_tacacs_server/tacacsserver_provider_nondefaults.rb
@@ -94,16 +94,16 @@ test_name "TestCase :: #{testheader}" do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
     cmd_str = PUPPET_BINPATH + 'resource cisco_tacacs_server'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout,
-                               { 'ensure'              => 'present',
-                                 'timeout'             => '50',
-                                 'deadtime'            => '0',
-                                 'encryption_password' => 'WXYZ12',
-                                 'directed_request'    => 'false',
-                                 'source_interface'    => 'Ethernet1/4' },
-                               false, self, logger)
-    end
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output,
+                             { 'ensure'              => 'present',
+                               'timeout'             => '50',
+                               'deadtime'            => '0',
+                               'encryption_password' => add_quotes('WXYZ12'),
+                               'directed_request'    => 'false',
+                               'source_interface'    => 'Ethernet1/4' },
+                             false, self, logger)
 
     logger.info("Check cisco_tacacs_server presence on agent :: #{result}")
   end
@@ -125,16 +125,16 @@ test_name "TestCase :: #{testheader}" do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to true to check for absence of RegExp pattern in stdout.
     cmd_str = PUPPET_BINPATH + 'resource cisco_tacacs_server'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout,
-                               { 'ensure'              => 'present',
-                                 'timeout'             => '50',
-                                 'deadtime'            => '0',
-                                 'encryption_password' => 'WXYZ12',
-                                 'directed_request'    => 'false',
-                                 'source_interface'    => 'Ethernet1/4' },
-                               true, self, logger)
-    end
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output,
+                             { 'ensure'              => 'present',
+                               'timeout'             => '50',
+                               'deadtime'            => '0',
+                               'encryption_password' => add_quotes('WXYZ12'),
+                               'directed_request'    => 'false',
+                               'source_interface'    => 'Ethernet1/4' },
+                             true, self, logger)
 
     logger.info("Check cisco_tacacs_server absence on agent :: #{result}")
   end