+++ tests/beaker_tests/radius_server/radius_server_provider_defaults.rb
@@ -83,26 +83,26 @@ test_name "TestCase :: #{testheader}" do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
     cmd_str = PUPPET_BINPATH + 'resource radius_server 8.8.8.8'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'accounting_only' => 'true' },
-                               false, self, logger) unless operating_system == 'ios_xr'
-      search_pattern_in_output(stdout, { 'acct_port' => '66' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'auth_port' => '77' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'authentication_only' => 'true' },
-                               false, self, logger) unless operating_system == 'ios_xr'
-      search_pattern_in_output(stdout, { 'key' => '44444444' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'key_format' => '7' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'retransmit_count' => '4' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'timeout' => '2' },
-                               false, self, logger)
-    end
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output, { 'ensure' => 'present' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'accounting_only' => 'true' },
+                             false, self, logger) unless operating_system == 'ios_xr'
+    search_pattern_in_output(output, { 'acct_port' => '66' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'auth_port' => '77' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'authentication_only' => 'true' },
+                             false, self, logger) unless operating_system == 'ios_xr'
+    search_pattern_in_output(output, { 'key' => add_quotes('44444444') },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'key_format' => '7' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'retransmit_count' => '4' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'timeout' => '2' },
+                             false, self, logger)
 
     logger.info("Check radius_server resource presence on agent :: #{result}")
   end
@@ -190,26 +190,26 @@ test_name "TestCase :: #{testheader}" do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
     cmd_str = PUPPET_BINPATH + 'resource radius_server 2003::7'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'accounting_only' => 'true' },
-                               false, self, logger) unless operating_system == 'ios_xr'
-      search_pattern_in_output(stdout, { 'acct_port' => '66' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'auth_port' => '77' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'authentication_only' => 'true' },
-                               false, self, logger) unless operating_system == 'ios_xr'
-      search_pattern_in_output(stdout, { 'key' => '44444444' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'key_format' => '7' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'retransmit_count' => '4' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'timeout' => '2' },
-                               false, self, logger)
-    end
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output, { 'ensure' => 'present' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'accounting_only' => 'true' },
+                             false, self, logger) unless operating_system == 'ios_xr'
+    search_pattern_in_output(output, { 'acct_port' => '66' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'auth_port' => '77' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'authentication_only' => 'true' },
+                             false, self, logger) unless operating_system == 'ios_xr'
+    search_pattern_in_output(output, { 'key' => add_quotes('44444444') },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'key_format' => '7' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'retransmit_count' => '4' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'timeout' => '2' },
+                             false, self, logger)
 
     logger.info("Check radius_server resource presence on agent :: #{result}")
   end