+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.5.0] - 2017-01-11
+
+### New feature support
+#### Cisco Resources
+- `cisco_hsrp_global` type and provider.
+- `cisco_interface_hsrp_group` type and provider.
+
+### Added
+- Extend cisco_interface with attributes:
+ - `hsrp_bfd`
+ - `hsrp_delay_minimum`
+ - `hsrp_delay_reload`
+ - `hsrp_mac_refresh`
+ - `hsrp_use_bia`
+ - `hsrp_version`
+ - `pim_bfd`
+
+- Extend cisco_pim with attributes:
+ - `bfd`
+
+* Added support for Cisco NX-OS software releases `7.3(0)F1(1)` and `8.0(1)`
+
+### Resolved Issues
+* https://github.com/cisco/cisco-network-puppet-module/issues/408
+
 ## [1.4.1] - 2016-11-02
 
 ### Added
@@ -312,6 +337,7 @@ This version was never released.
 - Initial release of puppetlabs-ciscopuppet module, supporting Cisco NX-OS software release 7.0(3)I2(1) on Cisco Nexus switch platforms: N95xx, N93xx, N30xx and N31xx.
 - Please note: 0.9.0 is an EFT pre-release for a limited audience with access to NX-OS 7.0(3)I2(1). Additional code changes may occur in 0.9.x prior to the final 1.0.0 release.
 
+[1.5.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.4.1...v1.5.0
 [1.4.1]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.3.1...v1.3.2